feat: add sunburst view for assets
