+++ packages/devtools/src/app/state/settings.ts
@@ -13,7 +13,7 @@ export interface ClientSettings {
   flowShowAllTransforms: boolean
   flowShowAllLoads: boolean
   moduleGraphViewType: 'list' | 'graph' | 'folder'
-  assetViewType: 'list' | 'folder' | 'treemap'
+  assetViewType: 'list' | 'folder' | 'treemap' | 'sunburst'
   chartAnimation: boolean
   moduleDetailsViewType: 'flow' | 'charts' | 'imports'
 }
